Release version 0.1.6
This project started with a very basic requirement i.e. creating a React
app with a very simple UI that just fetches the data from API and renders
it and should be easy on eyes. But, as I started working on it. I started
adding more and more features, and creating more requirements. At this
time I think the project is starting to get big enough to warrant a
version control.